Extract quick diff widget into separate file
